// packages/core/src/features/toast/enums.ts
export enum ToastKind {
  LegacySimulationAccess = "LegacySimulationAccess",
  None = "None",
  ProjectEditable = "ProjectEditable",
  ProjectForked = "ProjectForked",
  ProjectPreview = "ProjectPreview",
  ReadOnlyRelease = "ReadOnlyRelease",
  ReleaseBehaviorSuccess = "ReleaseBehaviorSuccess",
  ReleaseSuccess = "ReleaseSuccess",
}